database/__init__.py
import imdb
from .database import (
    add_connection,
    all_connections,
    if_active,
    delete_connection,
    make_active,
    make_inactive,
    del_all,
    find_filter,
    add_filter,
    find_filter,
    get_filters,
    delete_filter,
    count_filters,
    active_connection,
    add_user,
    all_users, 
    parser,
    split_quotes,
    donlee_imdb,
    Database,
    send_msg,
    add_user,   
    find_user,
    filter_stats, 
    humanbytes,
    google_search,
    remove_emoji
)

IMDBCONTROL = imdb.IMDb()
