requirements.txt
Dnspython
Motor
Pyrogram
Pymongo
Requests 
TgCrypto 
IMDbPY
asyncio
aiofiles
umongo
dnspython
heroku3
pyromod
